+++ acme/protocols/WebSocketServer.py
@@ -23,7 +23,7 @@ from ..helpers.ThreadSafeCounter import ThreadSafeCounter
 from ..etc.RequestUtils import prepareResultForSending, createPositiveResponseResult, createRequestResultFromURI
 from ..etc.IDUtils import uniqueID, csiFromSPRelative
 from ..etc.Utils import renameThread
-from ..etc.Types import ContentSerializationType, Result, CSERequest, Operation, ResourceTypes, RequestType, ResponseType
+from ..etc.Types import ContentSerializationType, Result, CSERequest, Operation, ResourceTypes, RequestType, ResponseType, AuthorizationResult
 from ..etc.ResponseStatusCodes import ResponseStatusCode, ResponseException, TARGET_NOT_REACHABLE, ORIGINATOR_HAS_NO_PRIVILEGE
 from ..runtime.Configuration import Configuration
 from ..runtime import CSE
@@ -334,7 +334,7 @@ class WebSocketServer(object):
 		return True
 
 
-	def receiveLoop(self, websocket:WSConnection, wsOriginator:str, ct:ContentSerializationType) -> None:
+	def receiveLoop(self, websocket:WSConnection, wsOriginator:str, ct:ContentSerializationType, authResult:AuthorizationResult) -> None:
 		"""	Receive loop for the WebSocket server. This is the main entry point for handling a received message,
 			whether the connection was initiated by the server or the client.
 
@@ -342,6 +342,7 @@ class WebSocketServer(object):
 				websocket: The WebSocket connection.
 				wsOriginator: The originator of the connection.
 				ct: The content type.
+				authResult: The result of the request authentication.
 		"""
 		try:	
 			# Handle incoming requests in separate threads as long as there is no error or the server is stopped
@@ -351,7 +352,7 @@ class WebSocketServer(object):
 				if not self._checkIsServerRunning(websocket):
 					continue
 				# Run the message handling in a separate thread
-				BackgroundWorkerPool.runJob(lambda: self._handleReceivedMessage(websocket, message, wsOriginator, ct), name = f'ws_{uniqueID()}')
+				BackgroundWorkerPool.runJob(lambda: self._handleReceivedMessage(websocket, message, wsOriginator, ct, authResult), name = f'ws_{uniqueID()}')
 		except ConnectionClosedError as e:
 			L.isWarn and L.logWarn('Connection closed: {e}')
 		except ConnectionClosedOK:
@@ -384,7 +385,7 @@ class WebSocketServer(object):
 		L.logDebug(f'Received request: {websocket.request}')
 
 		# Check the authentication
-		if not self._handleAuthentication(websocket):
+		if (authResult := self._handleAuthentication(websocket)) == AuthorizationResult.UNAUTHORIZED:
 			raise ORIGINATOR_HAS_NO_PRIVILEGE(L.logWarn('Authorization failed'))
 
 
@@ -413,7 +414,11 @@ class WebSocketServer(object):
 				if not self._checkIsServerRunning(websocket):
 					continue
 				# Run the message handling in a separate thread
-				BackgroundWorkerPool.runJob(lambda: self._handleReceivedMessage(websocket, message, wsOriginator, contentType),
+				BackgroundWorkerPool.runJob(lambda: self._handleReceivedMessage(websocket, 
+																				message, 
+																				wsOriginator, 
+																				contentType, 
+																				authResult),
 											name = f'ws_{uniqueID()}')
 		except ConnectionClosedError as e:
 			L.isWarn and L.logWarn('Connection closed: {e}')
@@ -424,7 +429,11 @@ class WebSocketServer(object):
 		self.removeConnection(websocket, wsOriginator)
 
 
-	def _handleReceivedMessage(self, websocket:WSConnection, message:str|bytes, wsOriginator:str, contentType:ContentSerializationType) -> None:
+	def _handleReceivedMessage(self, websocket:WSConnection, 
+									 message:str|bytes, 
+									 wsOriginator:str, 
+									 contentType:ContentSerializationType,
+									 authResult:AuthorizationResult) -> None:
 		"""	Handle a received message. This is the main entry point for handling a received message, whether the
 			message is a request or a response.
 
@@ -433,6 +442,7 @@ class WebSocketServer(object):
 				message: The received message.
 				wsOriginator: The originator of the connection.
 				contentType: The content type.
+				authResult: The result of the request authentication.
 		"""
 		if isinstance(message, str):
 			message = message.encode()	# Encode to bytes
@@ -451,6 +461,9 @@ class WebSocketServer(object):
 
 			request = dissectResult.request	# type:ignore [attr-defined]
 			requestOriginator:str = request.originator	# type:ignore [attr-defined]
+
+			# Add Authorization result to the request
+			request.rq_authn = authResult == AuthorizationResult.AUTHORIZED
 
 			# Check whether the message is a response or a request. If it is a response, then put it into the
 			# response queue and return. Another process might have sent the request and is waiting for the response.
@@ -486,6 +499,7 @@ class WebSocketServer(object):
 
 			L.isDebug and L.logDebug(f'Operation: {request.op}')
 			L.isDebug and L.logDebug(f'Originator: {requestOriginator}')
+			L.isDebug and L.logDebug(f'Authorization: {authResult}')
 
 			responseResult = CSE.request.handleRequest(request)
 
@@ -520,14 +534,14 @@ class WebSocketServer(object):
 		websocket.send(_data)
 
 
-	def _handleAuthentication(self, websocket:WSConnection) -> bool:
+	def _handleAuthentication(self, websocket:WSConnection) -> AuthorizationResult:
 		"""	Handle the authentication of a new connection.
 
 			Args:
 				websocket: The WebSocket connection.
 
 			Returns:
-				True if the authentication was successful, False otherwise.
+				Eneumeration value of the result of the authentication.
 		"""
 
 		def testBasicAuthentication(auth:str) -> bool:
@@ -567,20 +581,19 @@ class WebSocketServer(object):
 		if not (Configuration.websocket_security_enableBasicAuth or Configuration.websocket_security_enableTokenAuth):
 			if authorization is not None:
 				L.isWarn and L.logWarn('Basic or token authentication is not enabled, but an authorization header was found.')
-			return True
+			return AuthorizationResult.NOTSET
 		
 		if authorization is None:
 			L.isDebug and L.logDebug('No authorization header found.')
-			return False
+			return AuthorizationResult.UNAUTHORIZED
 		
 		if authorization.startswith('Basic '):
-			return testBasicAuthentication(authorization[6:])
+			return AuthorizationResult.AUTHORIZED if testBasicAuthentication(authorization[6:]) else AuthorizationResult.UNAUTHORIZED
 		elif authorization.startswith('Bearer '):
-			return testTokenAuthentication(authorization[7:])
+			return AuthorizationResult.AUTHORIZED if testTokenAuthentication(authorization[7:]) else AuthorizationResult.UNAUTHORIZED
 		else:
 			L.isWarn and L.logWarn(f'Unsupported authentication method: {authorization}')
-			return False
-		
+			return AuthorizationResult.UNAUTHORIZED
 
 
 	def sendWSRequest(self, request:CSERequest, url:str, ignoreResponse:bool) -> Result:
@@ -625,13 +638,16 @@ class WebSocketServer(object):
 
 			# Construct addional headers
 			additionalHeaders = { 'X-m2m-Origin': RC.cseCsi }	#  Always add the originator
+			authResult = AuthorizationResult.NOTSET
 			if request.credentials:	# Add the credentials if available
 				if request.credentials.wsUsername and request.credentials.wsPassword:
 					additionalHeaders['Authorization'] = request.credentials.getWsBasic()
+					authResult = AuthorizationResult.AUTHORIZED	# Assume success. Otherwise, the connection would not be established anyway
 				elif request.credentials.wsToken:
 					additionalHeaders['Authorization'] = request.credentials.getWsBearerToken()
+					authResult = AuthorizationResult.AUTHORIZED # Assume success. Otherwise, the connection would not be established anyway
 				else:
-					L.logWarn('No credentials found for WS request found')
+					L.logWarn('No credentials for WS request found')
 
 
 			# Else connect to the target WS server using the URL
@@ -647,7 +663,7 @@ class WebSocketServer(object):
 			self.addConnection(websocket, True)
 			self.associateConnectionWithOriginator(websocket, target) # In this case, the target is the originator
 
-			BackgroundWorkerPool.runJob(lambda: self.receiveLoop(websocket, target, ct),
+			BackgroundWorkerPool.runJob(lambda: self.receiveLoop(websocket, target, ct, authResult),
 										name = f'ws_{uniqueID()}')
 			return websocket, True